This file is a publication related to this code
Difference reside in the experimentation, consists on to execute a classifier algorithm to predict the corresponding class for each tweet for entities. Each experiment is carried out by removing the tweets of the entity to be evaluated and then, the rest of the tweets are used for as training corpus, from which the vocabulary is learned.